Update PIP_Manager.cpp
Successfully programmed and fixed position of the buttons. Got the program to run a pre-check of installed python packages and see which ones are outdated (Need to store it later so the start-up is quicker). 
Added a new tab that displays currently installed packages with the ones that are outdated (highlighted in red).
Added a search feature to the installed packages tab (Just type away and it'll search!)

There are a lot of comments, which mostly are pieces of code I've saved for later or temporarily scrapped. I will go back and comment out my code later on.